requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
clarifai==2.6.2
configparser==3.8.1
Django==3.0.5
future==0.18.2
googleapis-common-protos==1.51.0
grpcio==1.28.1
idna==2.9
isort==4.3.21
jsonschema==2.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.11.3
pylint==2.4.4
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.11.2
